+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 0.3.2 - 2023-06-20
 ### Added
 - Added new scenario status `aborted`.
 - Added endpoint `PUT /api/scenarios/:scenarioId/abort`. Scenario can be aborted until it is finished (`completed` or `failed`).